requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
sqlalchemy==1.4.23
python-multipart==0.0.5
Jinja2==3.0.1
aiofiles==0.7.0
opencv-python==4.8.0.74
numpy==1.26.2
Pillow==11.0.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==3.2.0
python-dotenv==0.19.0 